+++ backend/src/middlewares/index.ts
@@ -2,3 +2,4 @@ export * from "./error-middleware.js";
 export * from "./not-found-middleware.js";
 export * from "./rate-limit.js";
 export * from "./auth-middleware.js";
+export * from "./request-logger.middleware.js";